refactor: agregar metadatos de categorías de productos y reorganizar componentes de filtros
